backend/modules/IntentAnticipation/requirements.txt
flask
flask-cors
ollama
python-dotenv
rdflib
pandas
pykeen